recoding solution to have .cpp files for a clearer project.
