+++ Hooks/InternetOpenUrlW.h
@@ -1,5 +1,6 @@
 #pragma once
 
+
 typedef HINTERNET(WINAPI* InternetOpenUrlW_t)(HINTERNET hInternet, LPCWSTR lpszUrl, LPCWSTR lpszHeaders, DWORD dwHeadersLength, DWORD dwFlags, DWORD_PTR dwContext);
 
 InternetOpenUrlW_t oInternetOpenUrlW;
