+++ Hooks/NtCreateThreadEx.h
@@ -1,10 +1,10 @@
 #pragma once
 
 typedef NTSTATUS(NTAPI* NtCreateThreadEx_t)(OUT PHANDLE ThreadHandle, IN ACCESS_MASK DesiredAccess,
-    IN POBJECT_ATTRIBUTES ObjectAttributes, IN HANDLE ProcessHandle,
-    IN LPTHREAD_START_ROUTINE StartRoutine, IN LPVOID Argument,
-    IN ULONG CreateFlags, IN SIZE_T StackZeroBits, IN SIZE_T SizeOfStackCommit,
-    IN SIZE_T SizeOfStackReserve, OUT LPVOID BytesBuffer);
+                                            IN POBJECT_ATTRIBUTES ObjectAttributes, IN HANDLE ProcessHandle,
+                                            IN LPTHREAD_START_ROUTINE StartRoutine, IN LPVOID Argument,
+                                            IN ULONG CreateFlags, IN SIZE_T StackZeroBits, IN SIZE_T SizeOfStackCommit,
+                                            IN SIZE_T SizeOfStackReserve, OUT LPVOID BytesBuffer);
 
 NtCreateThreadEx_t oNtCreateThreadEx;
