+++ Includes/Includes.h
@@ -9,12 +9,16 @@
 #include <string>
 #include <vector>
 #include <dbghelp.h>
+#include <unordered_map>
+#include <codecvt>
+#include <locale>
+#include <iostream>
 #pragma comment(lib,"dbghelp.lib")
 
 #include <d3d9.h>
 #pragma comment(lib, "d3dx9")
 #pragma comment(lib,"d3d9.lib")
 
-#include "ImGui/imgui.h"
-#include "ImGui/imgui_impl_win32.h"
-#include "ImGui/imgui_impl_dx9.h"
+#include <imgui.h>
+#include <imgui_impl_win32.h>
+#include <imgui_impl_dx9.h>
