+++ ReverseKit.cpp
@@ -1,12 +1,8 @@
-#include "ReverseLib/Includes.h"
-#include "Instrumentation/InstrumentationCallback.h"
-#include "Window/D3DWindow.h"
-#include "Window/ImGuiSetup.h"
-#include "Imports/Imports.h"
-#include "Threads/Threads.h"
-#include "Hooks/SetHooks.h"
-#include "Menu/Menu.h"
-#include "Window/Render.h"
+#include <Includes.h>
+#include <Imports.h>
+#include <Threads.h>
+#include <SetHooks.h>
+#include <Render.h>
 
 
 DWORD WINAPI RetrievalThread(LPVOID lpParameter) 
@@ -18,11 +14,14 @@ DWORD WINAPI RetrievalThread(LPVOID lpParameter)
     }
     */
 
+    
     while (true) {
+        std::cout << "retrieving...\n";
         GetImportsFromIAT();
         GetThreadInformation();
         Sleep(5000);
     }
+    return TRUE;
 }
 
 BOOL APIENTRY DllMain(HMODULE hModule, DWORD  ul_reason_for_call, LPVOID lpReserved) 
@@ -31,9 +30,15 @@ BOOL APIENTRY DllMain(HMODULE hModule, DWORD  ul_reason_for_call, LPVOID lpReser
     {
     case DLL_PROCESS_ATTACH:
         SetConsoleTitleA("ReverseKit Attached");
+    	AllocConsole();
+		freopen("CONIN$", "r", stdin);
+		freopen("CONOUT$", "w", stderr);
+        freopen("CONOUT$", "w", stdout);
+        
         HookSyscalls();
-        CreateThread(nullptr, 0, RetrievalThread, nullptr, 0, nullptr);
-        CreateThread(nullptr, 0, RenderThread, nullptr, 0, nullptr);
+        
+        CreateThread(nullptr, 0, RetrievalThread, hModule, 0, nullptr);
+        CreateThread(nullptr, 0, RenderThread, hModule, 0, nullptr);
         break;
     case DLL_PROCESS_DETACH:
         UnhookSyscalls();
