+++ Window/Render.cpp
@@ -1,0 +1,72 @@
+#include "Render.h"
+
+#include "D3DWindow.h"
+#include "Menu.h"
+
+DWORD WINAPI RenderThread(LPVOID lpParameter)
+{
+    D3DWindow window = D3DWindow();
+    window.CreateOverlayWindow();
+
+    ImGui::CreateContext();
+    ImGuiIO& io = ImGui::GetIO();
+    io.IniFilename = nullptr;
+    io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable;
+
+    ImGui::StyleColorsDark();
+
+    ImGui_ImplWin32_Init(window.main_hwnd);
+    ImGui_ImplDX9_Init(window.g_pd3dDevice);
+
+    MSG msg;
+    ZeroMemory(&msg, sizeof(msg));
+
+    while (msg.message != WM_QUIT) {
+
+        if (PeekMessage(&msg, NULL, 0U, 0U, PM_REMOVE)) {
+            TranslateMessage(&msg);
+            DispatchMessage(&msg);
+            continue;
+        }
+
+        ImGui_ImplDX9_NewFrame();
+        ImGui_ImplWin32_NewFrame();
+        ImGui::NewFrame();
+
+        RenderUI();
+
+        ImGui::EndFrame();
+        window.g_pd3dDevice->Clear(0, NULL, D3DCLEAR_TARGET | D3DCLEAR_ZBUFFER, 0, 1.0f, 0);
+
+        if (window.g_pd3dDevice->BeginScene() >= 0)
+        {
+            ImGui::Render();
+            ImGui_ImplDX9_RenderDrawData(ImGui::GetDrawData());
+            window.g_pd3dDevice->EndScene();
+        }
+
+        if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable)
+        {
+            ImGui::UpdatePlatformWindows();
+            ImGui::RenderPlatformWindowsDefault();
+        }
+
+        const HRESULT result = window.g_pd3dDevice->Present(NULL, NULL, NULL, NULL);
+        if (result == D3DERR_DEVICELOST && window.g_pd3dDevice->TestCooperativeLevel() == D3DERR_DEVICENOTRESET) {
+            ImGui_ImplDX9_InvalidateDeviceObjects();
+        	const HRESULT hr = window.g_pd3dDevice->Reset(&window.g_d3dpp);
+			if (hr == D3DERR_INVALIDCALL)
+				IM_ASSERT(0);
+            ImGui_ImplDX9_CreateDeviceObjects();
+        }
+
+    }
+
+    ImGui_ImplDX9_Shutdown();
+    ImGui_ImplWin32_Shutdown();
+    ImGui::DestroyContext();
+
+    window.CleanupD3DDevice();
+
+    return 0;
+}
