+++ Window/Render.h
@@ -1,64 +1,5 @@
 #pragma once
+#include <Includes.h>
 
-DWORD WINAPI RenderThread(LPVOID lpParameter)
-{
-    CreateOverlayWindow();
 
-    ImGui::CreateContext();
-    ImGuiIO& io = ImGui::GetIO();
-    io.IniFilename = nullptr;
-    io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable;
-
-    ImGui::StyleColorsDark();
-
-    ImGui_ImplWin32_Init(main_hwnd);
-    ImGui_ImplDX9_Init(g_pd3dDevice);
-
-    MSG msg;
-    ZeroMemory(&msg, sizeof(msg));
-
-    while (msg.message != WM_QUIT) {
-
-        if (PeekMessage(&msg, NULL, 0U, 0U, PM_REMOVE)) {
-            TranslateMessage(&msg);
-            DispatchMessage(&msg);
-            continue;
-        }
-
-        ImGui_ImplDX9_NewFrame();
-        ImGui_ImplWin32_NewFrame();
-        ImGui::NewFrame();
-
-        RenderUI();
-
-        ImGui::EndFrame();
-        g_pd3dDevice->Clear(0, NULL, D3DCLEAR_TARGET | D3DCLEAR_ZBUFFER, 0, 1.0f, 0);
-
-        if (g_pd3dDevice->BeginScene() >= 0)
-        {
-            ImGui::Render();
-            ImGui_ImplDX9_RenderDrawData(ImGui::GetDrawData());
-            g_pd3dDevice->EndScene();
-        }
-
-        if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable)
-        {
-            ImGui::UpdatePlatformWindows();
-            ImGui::RenderPlatformWindowsDefault();
-        }
-
-        HRESULT result = g_pd3dDevice->Present(NULL, NULL, NULL, NULL);
-        if (result == D3DERR_DEVICELOST && g_pd3dDevice->TestCooperativeLevel() == D3DERR_DEVICENOTRESET) {
-            ResetD3DDevice();
-        }
-
-    }
-
-    ImGui_ImplDX9_Shutdown();
-    ImGui_ImplWin32_Shutdown();
-    ImGui::DestroyContext();
-
-    CleanupD3DDevice();
-
-    return 0;
-}
+DWORD WINAPI RenderThread(LPVOID lpParameter);
